Namespace ViewSpecHelper to Spec::Support::ViewHelper
