+++ spec/support/view_helper.rb
@@ -1,0 +1,36 @@
+require 'hamlit-rails'
+
+module Spec
+  module Support
+    module ViewHelper
+      def show_element(resp, id)
+        set_element_visibility(resp, id, true)
+      end
+
+      def hide_element(resp, id)
+        set_element_visibility(resp, id, false)
+      end
+
+      private
+
+      def set_element_visibility(resp, id, visible)
+        c = Capybara.string(resp)
+        node = c.find(id, :visible => :all)
+        node.native["style"] = node.native["style"].gsub(/display\s*:[^;]+/, "display:#{visible ? "show" : "none"}")
+        c.native.to_s
+      end
+
+      def set_controller_for_view(controller_name)
+        controller.request.path_parameters[:controller] = controller_name
+      end
+
+      def set_controller_for_view_to_be_restful
+        allow(controller).to receive(:restful?).and_return(true)
+      end
+
+      def set_controller_for_view_to_be_nonrestful
+        allow(controller).to receive(:restful?).and_return(false)
+      end
+    end
+  end
+end
